refactor(upsert-now): updated the requirements for the function you pass to XUpsertNow to return the
see the basic equality query for an example
